Move all tracker type to tracker service
