+++ services/tracker/pyproject.toml
@@ -14,11 +14,7 @@ dependencies = [
     "daytona>=0.128.1",
     "sqlmodel>=0.0.31",
     "httpx>=0.28.1",
-    "agentic-harness",
 ]
-
-[tool.uv.sources]
-agentic-harness = { path = "../.." }
 
 [dependency-groups]
 dev = [
@@ -33,7 +29,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/tracker"]
+packages=["src/tracker"]
 
 [tool.ruff]
 line-length = 120
