+++ services/tracker/tests/integration/test_sandbox.py
@@ -7,12 +7,12 @@ from unittest.mock import MagicMock, patch
 
 import boto3
 import pytest
-from agentic_harness import AgentContract
 from daytona import AsyncDaytona, AsyncSandbox, DaytonaError
 from moto import mock_aws
 from mypy_boto3_s3.client import S3Client
 
 from tracker.config import S3_BUCKET_NAME
+from tracker.types import AgentContract
 from tracker.s3 import get_contract_s3_key
 from tracker.sandbox import (
     create_sandbox,
