+++ src/agentic_harness/__init__.py
@@ -1,3 +1,3 @@
-from agentic_harness.contract import AgentContract
+from tracker.types import AgentContract
 
 __all__ = ["AgentContract"]
